# Use prepared statements for get/add transaction and moving temp