--- a/src/relstorage/adapters/connmanager.py
+++ b/src/relstorage/adapters/connmanager.py
@@ -63,6 +63,8 @@ class AbstractConnectionManager(object):
         Add a callable(cursor, restart=bool) for when a store connection
         is opened.
 
+        Hooks are called in the order added.
+
         .. versionadded:: 2.1a1
         """
         self._on_store_opened += (f,)
@@ -72,6 +74,8 @@ class AbstractConnectionManager(object):
     def add_on_load_opened(self, f):
         """
         Add a callable (cursor, restart=bool) for when a load connection is opened.
+
+        Hooks are called in the order added.
 
         .. versionadded:: 2.1a1
         """