--- a/src/relstorage/adapters/mysql/txncontrol.py
+++ b/src/relstorage/adapters/mysql/txncontrol.py
@@ -19,4 +19,7 @@ from ..txncontrol import GenericTransactionControl
 
 
 class MySQLTransactionControl(GenericTransactionControl):
-    pass
+
+    # See adapter.py for where this is prepared.
+    # Either history preserving or not, it's the same.
+    _get_tid_query = 'EXECUTE get_latest_tid'