--- a/src/relstorage/adapters/postgresql/mover.py
+++ b/src/relstorage/adapters/postgresql/mover.py
@@ -13,7 +13,7 @@
 ##############################################################################
 """IObjectMover implementation.
 """
-from __future__ import absolute_import
+from __future__ import absolute_import, print_function
 
 import functools
 import os
@@ -26,13 +26,47 @@ from ..interfaces import IObjectMover
 from ..mover import AbstractObjectMover
 from ..mover import metricmethod_sampled
 
+# Important: pg8000 1.10 - 1.13, at least, can't handle prepared
+# statements that take parameters but it doesn't need to because it
+# prepares every statement anyway. So you must have a backup that you use
+# for that driver.
+# https://github.com/mfenniak/pg8000/issues/132
 
-def to_prepared_queries(name, queries, datatypes=''):
-    # Only handles one param
-    return [
-        'PREPARE ' + name + ' ' + datatypes + ' AS ' + x.replace('%s', '$1')
-        for x in queries
-    ]
+
+def to_prepared_queries(name, queries, datatypes=()):
+    # Give correct datatypes for the queries, wherever possible.
+    # The number of parameters should be the same or more than the
+    # number of datatypes.
+    # datatypes is a sequence of strings.
+
+    # Maybe instead of having the adapter have to know about all the
+    # statements that need prepared, we could keep a registry?
+    if datatypes:
+        assert isinstance(datatypes, (list, tuple))
+        datatypes = ', '.join(datatypes)
+        datatypes = ' (%s)' % (datatypes,)
+    else:
+        datatypes = ''
+
+    result = []
+    for q in queries:
+        if not isinstance(q, str):
+            # Unsupported marker
+            result.append(q)
+            continue
+
+        q = q.strip()
+        param_count = q.count('%s')
+        rep_count = 0
+        while rep_count < param_count:
+            rep_count += 1
+            q = q.replace('%s', '$' + str(rep_count), 1)
+        stmt = 'PREPARE {name}{datatypes} AS {query}'.format(
+            name=name, datatypes=datatypes, query=q
+        )
+        result.append(stmt)
+    return result
+
 
 @implementer(IObjectMover)
 class PostgreSQLObjectMover(AbstractObjectMover):
@@ -42,7 +76,7 @@ class PostgreSQLObjectMover(AbstractObjectMover):
     _prepare_load_current_queries = to_prepared_queries(
         'load_current',
         AbstractObjectMover._load_current_queries,
-        '(BIGINT)')
+        ['BIGINT'])
 
     _prepare_load_current_query = query_property('_prepare_load_current')
 
@@ -62,6 +96,8 @@ class PostgreSQLObjectMover(AbstractObjectMover):
 
     # Sadly we can't PREPARE this statement; apparently it holds a
     # lock on OBJECT_STATE that interferes with taking the commit lock.
+    # XXX: Now that we've figured out to COMMIT right after the prepare,
+    # this should be fine and we should be able to prepare it.
     _move_from_temp_object_state_95_query = """
         INSERT INTO object_state (zoid, tid, state_size, state)
         SELECT zoid, %s, COALESCE(LENGTH(state), 0), state
@@ -79,40 +115,40 @@ class PostgreSQLObjectMover(AbstractObjectMover):
     def on_store_opened(self, cursor, restart=False):
         """Create the temporary tables for storing objects"""
         # note that the md5 column is not used if self.keep_history == False.
-        stmts = [
+        # Ideally we wouldn't execute any of these on a restart, but
+        # I've seen an issue with temp_stare apparently going missing on pg8000
+        ddl_stmts = [
             """
-            CREATE TEMPORARY TABLE temp_store (
-                zoid        BIGINT NOT NULL,
+            CREATE TEMPORARY TABLE IF NOT EXISTS temp_store (
+                zoid        BIGINT NOT NULL PRIMARY KEY,
                 prev_tid    BIGINT NOT NULL,
                 md5         CHAR(32),
                 state       BYTEA
-            ) ON COMMIT DROP;
+            ) ON COMMIT DELETE ROWS;
             """,
             """
-            CREATE UNIQUE INDEX temp_store_zoid ON temp_store (zoid);
-            """,
-            """
-            CREATE TEMPORARY TABLE temp_blob_chunk (
+            CREATE TEMPORARY TABLE IF NOT EXISTS temp_blob_chunk (
                 zoid        BIGINT NOT NULL,
                 chunk_num   BIGINT NOT NULL,
-                chunk       OID
-            ) ON COMMIT DROP;
-            """,
-            """
-            CREATE UNIQUE INDEX temp_blob_chunk_key
-            ON temp_blob_chunk (zoid, chunk_num);
-            """,
-            """
-            -- This trigger removes blobs that get replaced before being
-            -- moved to blob_chunk.  Note that it is never called when
-            -- the temp_blob_chunk table is being dropped or truncated.
-            CREATE TRIGGER temp_blob_chunk_delete
-                BEFORE DELETE ON temp_blob_chunk
-                FOR EACH ROW
-                EXECUTE PROCEDURE temp_blob_chunk_delete_trigger();
+                chunk       OID,
+                PRIMARY KEY (zoid, chunk_num)
+            ) ON COMMIT DELETE ROWS;
             """,
         ]
-        for stmt in stmts:
+        if not restart:
+            ddl_stmts += [
+                """
+                -- This trigger removes blobs that get replaced before being
+                -- moved to blob_chunk.  Note that it is never called when
+                -- the temp_blob_chunk table is being dropped or truncated.
+                CREATE TRIGGER temp_blob_chunk_delete
+                    BEFORE DELETE ON temp_blob_chunk
+                    FOR EACH ROW
+                    EXECUTE PROCEDURE temp_blob_chunk_delete_trigger();
+                """,
+            ]
+
+        for stmt in ddl_stmts:
             cursor.execute(stmt)
 
         if self.__need_version_check:
@@ -272,9 +308,6 @@ class PostgreSQLObjectMover(AbstractObjectMover):
 
 
 class PG8000ObjectMover(PostgreSQLObjectMover):
-    # pg8000 1.10 can't handle prepared statements that take parameters
-    # but it doesn't need to because it prepares every statement
-    # anyway. https://github.com/mfenniak/pg8000/issues/132
 
     on_load_opened_statement_names = ()
     on_store_opened_statement_names = ('_prepare_detect_conflict_query',)