These are typically called once per transaction. For larger transactions, no difference shows up in benchmarks. But for small ones, it's there:

| Benchmark                        | Master | This branch              |
|----------------------------------|-----------------|------------------------------|
| psycopg2_0: add 10 objects       | 9.42 ms                | 6.64 ms: 1.42x faster (-30%) |
| psycopg2_0: update 10 objects    | 10.6 ms                | 7.58 ms: 1.40x faster (-29%) |
| psycopg2_0_hf: add 10 objects    | 9.75 ms                | 5.17 ms: 1.89x faster (-47%) |
| psycopg2_0_hf: update 10 objects | 11.2 ms                | 6.47 ms: 1.73x faster (-42%) |


I'd like to continue moving towards using prepared statements where it makes sense, but this PR shows that the infrastructure to make that happen could use some work. Currently it's pretty ugly.

Fixes #246