@jamadden 2019-06-13
The above benchmarks were run with local caching turned off, to minimize the extra overheads outside what I was trying to measure. The effect is still there with the default cache settings, just smaller. Even without this change, though, the cache has gotten faster in 3.0a1 as compared to 2.1.1:

| Benchmark                        | 2.1.1               | Master                       | This branch                  |
|----------------------------------|---------------------|------------------------------|------------------------------|
| psycopg2_0: add 10 objects       | 11.9 ms             | 10.9 ms: 1.09x faster (-8%)  | 8.23 ms: 1.44x faster (-31%) |
| psycopg2_0: update 10 objects    | 12.3 ms             | 11.4 ms: 1.08x faster (-8%)  | 8.85 ms: 1.40x faster (-28%) |
| psycopg2_0_hf: add 10 objects    | 12.9 ms             | 11.5 ms: 1.12x faster (-11%) | 8.70 ms: 1.49x faster (-33%) |
| psycopg2_0_hf: update 10 objects | 13.1 ms             | 11.7 ms: 1.12x faster (-11%) | 9.17 ms: 1.43x faster (-30%) |